@@ -10,8 +10,8 @@ plugins {
     id 'signing'
 }
 
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
+sourceCompatibility = JavaVersion.VERSION_11
+targetCompatibility = JavaVersion.VERSION_11
 
 group = 'com.nulab-inc'
 archivesBaseName = 'backlog4j'
@@ -28,11 +28,12 @@ dependencies {
     testImplementation 'org.hamcrest:hamcrest-library:2.2'
     testImplementation 'org.exparity:hamcrest-date:2.0.8'
     testImplementation 'ch.qos.logback:logback-classic:1.2.11'
+    testImplementation 'com.squareup.okhttp3:mockwebserver:4.11.0'
 
     implementation 'org.apache.commons:commons-lang3:3.12.0'
-    implementation 'com.fasterxml.jackson.core:jackson-core:2.13.3'
-    implementation 'com.fasterxml.jackson.core:jackson-annotations:2.13.3'
-    implementation 'com.fasterxml.jackson.core:jackson-databind:2.13.3'
+    implementation 'com.fasterxml.jackson.core:jackson-core:2.15.3'
+    implementation 'com.fasterxml.jackson.core:jackson-annotations:2.15.3'
+    implementation 'com.fasterxml.jackson.core:jackson-databind:2.15.3'
     implementation 'org.slf4j:slf4j-api:1.7.36'
 }
 
@@ -82,7 +83,7 @@ publishing {
                 licenses {
                     license {
                         name = 'MIT License'
-                        url = 'http://www.opensource.org/licenses/mit-license.php'
+                        url = 'https://opensource.org/license/mit/'
                     }
                 }
                 developers {
