@@ -5,9 +5,9 @@ import com.nulabinc.backlog4j.auth.AccessToken;
 import com.nulabinc.backlog4j.auth.OAuthSupport;
 import com.nulabinc.backlog4j.conf.BacklogConfigure;
 import com.nulabinc.backlog4j.http.BacklogHttpClient;
-import com.nulabinc.backlog4j.http.BacklogHttpClientImpl;
 import com.nulabinc.backlog4j.http.BacklogHttpResponse;
 import com.nulabinc.backlog4j.http.NameValuePair;
+import com.nulabinc.backlog4j.http.httpclient.HttpClientBacklogHttpClient;
 import com.nulabinc.backlog4j.internal.InternalFactory;
 import com.nulabinc.backlog4j.internal.json.InternalFactoryJSONImpl;
 import org.slf4j.Logger;
@@ -35,7 +35,7 @@ public abstract class BacklogClientBase {
 
     public BacklogClientBase(BacklogConfigure configure) {
         this.configure = configure;
-        this.httpClient = new BacklogHttpClientImpl();
+        this.httpClient = new HttpClientBacklogHttpClient();
         this.backlogEndPointSupport = new BacklogEndPointSupport(configure);
         configureHttpClient();
     }
