@@ -9,10 +9,9 @@ public class BacklogClientImplTest {
     @Test
     public void userAgentTest() {
         BacklogHttpClient client = new BacklogHttpClientImpl();
-        String version = BacklogHttpClientImpl.class.getPackage().getImplementationVersion();
 
         // default
-        assertEquals(client.getUserAgent(), "backlog4j/" + version);
+        assertEquals(client.getUserAgent(), "backlog4j/-");
 
         // set the new value
         client.setUserAgent("test-agent");
