@@ -3,6 +3,8 @@ package com.nulabinc.backlog4j.internal.json;
 import java.io.CharArrayWriter;
 import java.io.IOException;
 import java.io.InputStreamReader;
+import java.nio.charset.StandardCharsets;
+import java.util.Objects;
 import java.util.TimeZone;
 
 /**
@@ -14,20 +16,14 @@ public abstract class AbstractJSONImplTest {
     protected TimeZone timeZone = TimeZone.getTimeZone("Asia/Tokyo");
 
     protected String getJsonString(String resourceFilePath) throws IOException {
-        InputStreamReader r = null;
-        try {
-            r = new InputStreamReader(Thread.currentThread().getContextClassLoader().getResourceAsStream(resourceFilePath), "UTF-8");
+        try (InputStreamReader r = new InputStreamReader(Objects.requireNonNull(Thread.currentThread().getContextClassLoader().getResourceAsStream(resourceFilePath)), StandardCharsets.UTF_8)) {
             CharArrayWriter writer = new CharArrayWriter();
-            char buff[] = new char[4096];
+            char[] buff = new char[4096];
             int size;
             while ((size = r.read(buff)) > 0) {
                 writer.write(buff, 0, size);
             }
             return writer.toString();
-        } finally {
-            if (r != null) {
-                r.close();
-            }
         }
     }
 }
